DTD-2858: Refactoring: Move & rename Identification/IDType/IDValue to make them consistent.
