+++ app/uk/gov/hmrc/timetopayproxy/models/chargeInfoApi/ChargeInfoRequest.scala
@@ -19,6 +19,7 @@ package uk.gov.hmrc.timetopayproxy.models.chargeInfoApi
 import cats.data.NonEmptyList
 import enumeratum.{ Enum, EnumEntry, PlayJsonEnum }
 import play.api.libs.json.{ Format, Json, OFormat }
+import uk.gov.hmrc.timetopayproxy.models.Identification
 import uk.gov.hmrc.timetopayproxy.utils.json.CatsNonEmptyListJson
 
 import scala.collection.immutable
