+++ test/uk/gov/hmrc/timetopayproxy/connectors/TtpeConnectorSpec.scala
@@ -26,7 +26,7 @@ import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.http.client.HttpClientV2
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 import uk.gov.hmrc.timetopayproxy.config.AppConfig
-import uk.gov.hmrc.timetopayproxy.models.ConnectorError
+import uk.gov.hmrc.timetopayproxy.models.{ ConnectorError, IdType, IdValue, Identification }
 import uk.gov.hmrc.timetopayproxy.models.TtppEnvelope.TtppEnvelope
 import uk.gov.hmrc.timetopayproxy.models.chargeInfoApi._
 import uk.gov.hmrc.timetopayproxy.support.WireMockUtils
@@ -95,8 +95,8 @@ class TtpeConnectorSpec
     def chargeInfoRequest: ChargeInfoRequest = ChargeInfoRequest(
       channelIdentifier = ChargeInfoChannelIdentifier("Channel Identifier"),
       identifications = NonEmptyList.of(
-        Identification(idType = IDType("id type 1"), idValue = IDValue("id value 1")),
-        Identification(idType = IDType("id type 2"), idValue = IDValue("id value 2"))
+        Identification(idType = IdType("id type 1"), idValue = IdValue("id value 1")),
+        Identification(idType = IdType("id type 2"), idValue = IdValue("id value 2"))
       ),
       regimeType = RegimeType.SA
     )
@@ -104,7 +104,7 @@ class TtpeConnectorSpec
     def chargeInfoResponse: ChargeInfoResponse = ChargeInfoResponse(
       processingDateTime = LocalDateTime.parse("2025-07-02T15:00:41.689"),
       identification = List(
-        Identification(idType = IDType("ID_TYPE"), idValue = IDValue("ID_VALUE"))
+        Identification(idType = IdType("ID_TYPE"), idValue = IdValue("ID_VALUE"))
       ),
       individualDetails = IndividualDetails(
         title = Some(Title("Mr")),
