+++ test/uk/gov/hmrc/timetopayproxy/models/chargeInfoApi/ChargeInfoRequestSpec.scala
@@ -20,6 +20,7 @@ import cats.data.NonEmptyList
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers._
 import play.api.libs.json.{ JsSuccess, JsValue, Json, Reads, Writes }
+import uk.gov.hmrc.timetopayproxy.models.{ IdType, IdValue, Identification }
 import uk.gov.hmrc.timetopayproxy.testutils.JsonAssertionOps._
 import uk.gov.hmrc.timetopayproxy.testutils.schematestutils.Validators
 
@@ -30,8 +31,8 @@ class ChargeInfoRequestSpec extends AnyFreeSpec {
       def obj: ChargeInfoRequest = ChargeInfoRequest(
         channelIdentifier = ChargeInfoChannelIdentifier("Channel Identifier"),
         identifications = NonEmptyList.of(
-          Identification(idType = IDType("id type 1"), idValue = IDValue("id value 1")),
-          Identification(idType = IDType("id type 2"), idValue = IDValue("id value 2"))
+          Identification(idType = IdType("id type 1"), idValue = IdValue("id value 1")),
+          Identification(idType = IdType("id type 2"), idValue = IdValue("id value 2"))
         ),
         regimeType = RegimeType.SA
       )
